feat: Pure Core(分類A)をplot-digitizer-coreへ移植し、既存側をre-exportラッパー化
docs/design/plot-digitizer-architecture.md Phase 1。分類A
(domain/models, axisSetCalculator, strategies, utils, magnifier, confirmer)
をpackages/plot-digitizer-coreへ移植し、既存アプリ(src/)は同名の
re-exportラッパーに置き換えた。既存の公開挙動・importパスは変更なし。

## 移植したコード

- domain/models: Axis, AxisSet, Dataset(+ Interface)
- domain/services: AxisSetCalculator
- application/strategies: ExtractParent, LineExtract, SymbolExtractByArea
  (+ Interface)
- application/utils: extractColorSwatches, getPointsTotalDistance
- application/services: Magnifier, Confirmer(+ Interface)

domain層の共有型/定数(Coord, Point, PointMode, POINT_MODE)は
packages/plot-digitizer-core/src/domain/{types,constants}.ts として
domain層内に配置(パッケージルート直下ではなく)。理由:
dependency-cruiserのcore-domain-no-outward-dependencyルールに抵触した
ため — 共有値型/定数はdomainが所有すべき語彙であり、
「domainより外側への依存」ではないと判断した。

## 既存テストの移植(TDD: Red→Green確認)

axis/axisSet/dataset/axisSetCalculator/lineExtract/symbolExtractByArea/
colorPaletteUtilsの既存テストをcoreへ移植。lineExtract/symbolExtractByArea
は.jsから.tsへ変換(Uint8ClampedArrayでラップ、配列は元の値と1対1で
一致することをスクリプトで検証)。移植後、アプリ側の重複テストは削除
(coreに一本化)。

さらに、コミット済みのカバレッジゲート(core全体90% / domain/models・
domain/services 95%)を満たすため、未テストだった分岐・関数を洗い出して
追加(magnifier/confirmer/pointsUtilsの新規テスト、axisSet/datasetの
未カバー分岐、isDrawnMaskパス、setter類など)。最終カバレッジ:
core全体 100% Stmts / 99.33% Branch / 98.51% Funcs / 100% Lines、
domain/models・domain/services は各100%。

## アプリ側の配線

- vite.config.js / vite.library.config.ts / 直下tsconfig.json / 直下
  jest.config.cjsに  パスエイリアスを追加
  (packages/plot-digitizer-core/src/index.tsへ直接解決。npm workspaces
  は不使用の既定方針のため)
- 直下jest.config.cjsにtestPathIgnorePatternsを追加し、
  packages/配下がアプリ側jestの二重実行対象にならないようにした
- 移植元だったファイルは全て同名のre-exportラッパーに置換
  (import元は一切変更不要)

## 検証(Haikuサブエージェントに委譲・2回実施)

- core: eslint / jest --coverage(124 tests) / tsc --noEmit
- app: eslint / vue-tsc --noEmit / jest(43 tests) / depcruise(0件) /
  vite build -c vite.library.config.ts(ライブラリビルド成功)
  すべて成功
