+++ packages/plot-digitizer-core/src/application/services/confirmer/confirmerInterface.ts
@@ -1,0 +1,16 @@
+export interface ConfirmerInterface {
+  isActive: Boolean
+  message: String
+  handleOnConfirm: Function
+  handleOnCancel: Function
+  activate({
+    message,
+    onConfirm,
+    onCancel,
+  }: {
+    message: String
+    onConfirm: Function
+    onCancel: Function
+  }): void
+  inactivate(): void
+}
