+++ packages/plot-digitizer-core/src/application/services/magnifier/magnifierInterface.ts
@@ -1,0 +1,10 @@
+export interface MagnifierInterface {
+  scale: number
+  magnifierSettingError: string
+  crosshairSizePx: number
+  sizePx: number
+  effectiveDigits: number
+
+  setScale(scale: number): void
+  setEffectiveDigits(digits: number): void
+}
