+++ packages/plot-digitizer-core/src/application/strategies/extractStrategies/extractStrategyInterface.ts
@@ -1,0 +1,15 @@
+import { Coord } from '../../../domain/types'
+
+// INFO: Strategy Pattern
+export default interface ExtractStrategyInterface {
+  name: string
+  execute(
+    height: number,
+    width: number,
+    imageColors: Uint8ClampedArray,
+    maskColors: Uint8ClampedArray,
+    isDrawnMask: boolean,
+    targetColor: [number, number, number],
+    colorMatchThreshold: number,
+  ): Coord[]
+}
