+++ src/application/services/magnifier/magnifier.ts
@@ -1,17 +1,6 @@
-import { MagnifierInterface } from './magnifierInterface'
-
-export class Magnifier implements MagnifierInterface {
-  scale = 5
-  magnifierSettingError = ''
-  crosshairSizePx = 1
-  sizePx = 300
-  effectiveDigits = 4
-
-  setScale(scale: number) {
-    this.scale = scale
-  }
-
-  setEffectiveDigits(digits: number) {
-    this.effectiveDigits = digits
-  }
-}
+// INFO: Phase 1 (docs/design/plot-digitizer-architecture.md) — this class
+// has been moved to packages/plot-digitizer-core. This file re-exports it so
+// every existing `@/application/services/magnifier/magnifier` (and relative
+// `../magnifier`) import in this app keeps working unchanged. Do not add
+// logic here — edit the core package instead.
+export { Magnifier } from '@plot-digitizer/core'
